# openapi.json
{
  "components": {
    "schemas": {
      "Ack": {
        "description": "Generic success envelope for endpoints with nothing meaningful to return.",
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "Ack",
        "type": "object"
      },
      "AgentState": {
        "enum": [
          "idle",
          "listening",
          "thinking",
          "speaking",
          "muted"
        ],
        "title": "AgentState",
        "type": "string"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code, e.g. `not_found`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context, e.g. field validation errors.",
            "title": "Detail"
          },
          "message": {
            "description": "Human-readable message. Safe to surface in UI.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Every non-2xx response from `/api` has this shape.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvictedBot": {
        "properties": {
          "bot_id": {
            "title": "Bot Id",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "left": {
            "description": "False when Recall refused the leave call.",
            "title": "Left",
            "type": "boolean"
          },
          "meeting_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Url"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "bot_id",
          "meeting_url",
          "status",
          "left",
          "detail"
        ],
        "title": "EvictedBot",
        "type": "object"
      },
      "EvictionReport": {
        "properties": {
          "checked": {
            "description": "Bots Recall reported as still in a call.",
            "title": "Checked",
            "type": "integer"
          },
          "evicted": {
            "items": {
              "$ref": "#/components/schemas/EvictedBot"
            },
            "title": "Evicted",
            "type": "array"
          }
        },
        "required": [
          "checked",
          "evicted"
        ],
        "title": "EvictionReport",
        "type": "object"
      },
      "Fixture": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "integer"
          },
          "id": {
            "examples": [
              "q3_revenue_review"
            ],
            "title": "Id",
            "type": "string"
          },
          "participant_count": {
            "title": "Participant Count",
            "type": "integer"
          },
          "title": {
            "examples": [
              "Q3 Revenue Review"
            ],
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "duration_seconds",
          "participant_count"
        ],
        "title": "Fixture",
        "type": "object"
      },
      "HarnessStart": {
        "properties": {
          "fixture_id": {
            "examples": [
              "q3_revenue_review"
            ],
            "title": "Fixture Id",
            "type": "string"
          },
          "loop": {
            "default": false,
            "description": "Restart the fixture when it ends.",
            "title": "Loop",
            "type": "boolean"
          },
          "speed": {
            "default": 1.0,
            "description": "Playback multiplier. Use 4-10 to iterate quickly; 1.0 to demo.",
            "maximum": 50.0,
            "minimum": 0.1,
            "title": "Speed",
            "type": "number"
          }
        },
        "required": [
          "fixture_id"
        ],
        "title": "HarnessStart",
        "type": "object"
      },
      "HarnessStop": {
        "properties": {
          "meeting_id": {
            "title": "Meeting Id",
            "type": "string"
          }
        },
        "required": [
          "meeting_id"
        ],
        "title": "HarnessStop",
        "type": "object"
      },
...
      "Meeting": {
        "properties": {
          "agent_state": {
            "$ref": "#/components/schemas/AgentState",
            "default": "idle"
          },
          "bot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Recall.ai bot id. Null when `source` is `harness`.",
            "title": "Bot Id"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "examples": [
              "mtg_01J8XM6N7P8Q9R0S1T2U3V"
            ],
            "title": "Id",
            "type": "string"
          },
          "meeting_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "https://meet.google.com/abc-defg-hij"
            ],
            "title": "Meeting Url"
          },
          "platform": {
            "$ref": "#/components/schemas/MeetingPlatform",
            "default": "google_meet"
          },
          "roster": {
            "items": {
              "$ref": "#/components/schemas/RosterEntry"
            },
            "title": "Roster",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/MeetingSource",
            "default": "recall"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/MeetingState"
          },
          "stats": {
            "$ref": "#/components/schemas/MeetingStats"
          },
          "title": {
            "examples": [
              "Q3 Revenue Review"
            ],
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "meeting_url",
          "platform",
          "state",
          "agent_state",
          "source",
          "bot_id",
          "roster",
          "started_at",
          "ended_at",
          "stats",
          "error"
        ],
        "title": "Meeting",
        "type": "object"
      },
...
      "MeetingPlatform": {
        "enum": [
          "google_meet",
          "zoom",
          "teams"
        ],
        "title": "MeetingPlatform",
        "type": "string"
      },
      "MeetingSource": {
        "enum": [
          "recall",
          "harness"
        ],
        "title": "MeetingSource",
        "type": "string"
      },
      "MeetingState": {
        "enum": [
          "scheduled",
          "joining",
          "in_call",
          "ended",
          "failed"
        ],
        "title": "MeetingState",
        "type": "string"
      },
...
      "Page_Fixture_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Fixture"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[Fixture]",
        "type": "object"
      },
...
      "RosterEntry": {
        "properties": {
          "display_name": {
            "examples": [
              "Sarah Chen"
            ],
            "title": "Display Name",
            "type": "string"
          },
          "is_host": {
            "default": false,
            "title": "Is Host",
            "type": "boolean"
          },
          "is_speaking": {
            "default": false,
            "title": "Is Speaking",
            "type": "boolean"
          },
          "matched": {
            "default": false,
            "description": "Whether this participant was resolved to a known Person. Unmatched participants should be visually flagged \u2014 Meet AGI has no context on them.",
            "title": "Matched",
            "type": "boolean"
          },
          "participant_id": {
            "description": "Platform-scoped, stable within the meeting.",
            "examples": [
              "p_2"
            ],
            "title": "Participant Id",
            "type": "string"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved Person, or null when unmatched.",
            "title": "Person Id"
          }
        },
        "required": [
          "participant_id",
          "person_id",
          "display_name",
          "is_host",
          "matched",
          "is_speaking"
        ],
        "title": "RosterEntry",
        "type": "object"
      },
...
      "TunnelStatus": {
        "description": "Whether Recall can actually reach this backend to deliver transcript.",
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "reachable": {
            "title": "Reachable",
            "type": "boolean"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Public webhook URL, token redacted.",
            "title": "Url"
          },
          "wake_word": {
            "title": "Wake Word",
            "type": "string"
          }
        },
        "required": [
          "configured",
          "reachable",
          "url",
          "wake_word",
          "detail"
        ],
        "title": "TunnelStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nBackend for **Meet AGI**, a Google Meet copilot that actually speaks.\n\n### Building the frontend against this\n\nYou do not need a Google Meet, a Recall.ai key, or an internet connection. Start a\nfixture replay and subscribe to the socket:\n\n```bash\ncurl -X POST localhost:8000/api/dev/harness/start \\\n  -H 'content-type: application/json' \\\n  -d '{\"fixture_id\":\"q3_revenue_review\",\"speed\":6,\"loop\":true}'\n```\n\nThat emits the complete real-time event stream \u2014 participants, partial and final\ntranscript, speaking states, wake-word detection, agent state transitions, and\ninterjections with real citations. It is byte-for-byte the same contract a real meeting\nproduces; only `Meeting.source` differs.\n\n### Contract rules\n\n- Types are generated from this OpenAPI document. Never hand-write an API type.\n- All mutations are REST. The WebSocket is server-to-client only.\n- Adding fields, enum variants, and event types is free and unannounced. Renames and\n  removals get a heads-up first.\n- `switch` on `LiveEvent.type` with a `default` branch. New event types will arrive.\n",
    "title": "Meet AGI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dev/evict-bots": {
      "post": {
        "description": "Asks Recall which bots are still in a call and tells each one to leave.\n\nA bot is a Recall-side entity: killing the backend that dispatched it does not remove it from the meeting, it just orphans it there until Recall's `everyone_left_timeout` expires. After a hard restart that leaves a silent, deaf Meet AGI sitting in the call next to the new one.\n\n`keep` spares a bot \u2014 normally the one the caller just dispatched.",
        "operationId": "evict_bots_api_dev_evict_bots_post",
        "parameters": [
          {
            "description": "Bot id to leave running.",
            "in": "query",
            "name": "keep",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bot id to leave running.",
              "title": "Keep"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvictionReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Remove stray bots from every meeting",
        "tags": [
          "dev"
        ]
      }
    },
    "/api/dev/fixtures": {
      "get": {
        "operationId": "list_fixtures_api_dev_fixtures_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Fixture_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "List available fixtures",
        "tags": [
          "dev"
        ]
      }
    },
    "/api/dev/harness/start": {
      "post": {
        "description": "Creates a meeting with `source: harness` and begins replaying. The event stream is identical to a real meeting \u2014 subscribe to `WS /api/meetings/{id}/live` as normal.\n\n`POST /api/meetings/{id}/speak` works on a harness meeting too. The audio goes nowhere, but the queueing, the real clip durations, and the `agent.state_changed` events are the same ones a live bot produces, so the speaking UI can be built against it.",
        "operationId": "start_harness_api_dev_harness_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Start a fixture replay",
        "tags": [
          "dev"
        ]
      }
    },
    "/api/dev/harness/stop": {
      "post": {
        "operationId": "stop_harness_api_dev_harness_stop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessStop"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Stop a fixture replay",
        "tags": [
          "dev"
        ]
      }
    },
    "/api/dev/reset": {
      "post": {
        "description": "Cancels running replays, clears meetings, transcripts, and interjections, and re-seeds the demo people, documents, and integrations. Also clears the pipeline's conversation memory and interjection cooldowns, and re-reads the `knowledge/` corpus from disk \u2014 so editing a document between demo runs takes effect without a restart. Handy between demo runs.",
        "operationId": "reset_api_dev_reset_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Reset all state",
        "tags": [
          "dev"
        ]
      }
    },
    "/api/dev/tunnel": {
      "get": {
        "description": "Calls this server's own health endpoint back through `PUBLIC_BASE_URL`. That round trip is the only thing that proves the tunnel is up \u2014 a configured URL with a dead tunnel is indistinguishable from a working one until somebody says the wake word into a meeting and nothing happens.",
        "operationId": "tunnel_status_api_dev_tunnel_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TunnelStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Can Recall reach us?",
        "tags": [
          "dev"
        ]
      }
    },
...
  }
}